creates the MovesTable component
